Add files via upload
A whimsical web project crafted with HTML, CSS, and JS. Experience the joy of  humor with a simple click. Press the button, and voilà! A new dad joke appears, ready to brighten your day. It's quick, it's fun, and it's your go-to source for instant laughter. Dad jokes, one click away!