Add files via upload
Experience form interaction like never before! With our Form Wave project, crafted with vanilla JavaScript, your input fields come to life. Click, and witness a distinctive CSS waving animation that adds a playful wavy effect to the text. It's not just a form; it's a dynamic and delightful user experience. Dive into the wave of creativity!