Add files via upload
H2OPtimize is a Vanilla JavaScript project that aims to revolutionize your hydration habits. This elegant water drinking guide combines the simplicity of Vanilla JavaScript with a visually appealing user interface crafted with CSS. The project is designed to promote a healthier lifestyle by encouraging users to stay adequately hydrated throughout the day.